Various bits of code cleanup (largely focused on cleaning up TODO messages or making them clearer).

+++ Kernel/Libraries/Kernel.Hardware/BlockDevice.cs
@@ -86,9 +86,7 @@ namespace Kernel.Hardware
         /// <returns>The new byte array.</returns>
         public byte[] NewBlockArray(UInt32 aBlockCount)
         {
-            //TODO - Err..this cast here is really really bad practice but it's just because 
-            //  we can't do 64 bit x 64 bit multiplication nor support the conv.ovf.i.un op
-            return new byte[aBlockCount * (UInt32)BlockSize];
+            return new byte[aBlockCount * BlockSize];
         }
     }
 }
